Apply `ptr_eq` lint only if `cmp_null` is not applicable
The `cmp_null` lint is more specialized than `ptr_eq`. The former should
take precedence, unless the user allows it.
